関数名等修正

+++ slack/run.py
@@ -6,7 +6,7 @@ from slackbot import settings
 from slackbot.bot import Bot
 import importlib
 
-def main():
+def start_bot():
     importlib.reload(slackbot_settings)
     importlib.reload(settings)
     bot = Bot()
@@ -21,28 +21,30 @@ class BotManager():
     def __init__(self):
         self.p = None
 
-
     def start(self):
         if not self.is_active():
-            self.p = multiprocessing.Process(target=main)
+            self.p = multiprocessing.Process(target=start_bot)
             self.p.start()
             print(self.p.is_alive(), os.getpid(), self.p.pid)
 
     def stop(self):
-        print('before terminate')
+        if not self.p.is_alive():
+            return
+
+        print('before termination')
         print(self.p.is_alive(), self.p.exitcode, self.p.pid)
-        print('try terminate')
+        print('try to terminate')
         self.p.terminate()
         result = self.p.join(0.5)
         print(self.p.is_alive(), self.p.exitcode, self.p.pid, result)
         if self.p.exitcode is None:
-            print('try terminate failed once')
-            print('try terminate again')
+            print('Termination failed once')
+            print('try to terminate again')
             os.system('kill -9 {}'.format(self.p.pid))
             result = self.p.join(0.5)
             print(self.p.is_alive(), self.p.exitcode, self.p.pid, result)
             if self.p.exitcode is None:
-                raise Exception('terminate process failed')
+                raise Exception('Termination process failed')
         if not self.p.is_alive():
             self.p = None
 
@@ -50,4 +52,4 @@ class BotManager():
         return self.p is not None
 
 if __name__ == '__main__':
-    main()
+    start_bot()
